Week_4/07102025/Week03_test_solution/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.25
pydantic>=2.0.0
PyJWT>=2.8.0
pytest>=8.0.0
httpx>=0.26.0
pydantic[email]
